requirements.txt
streamlit
python-dotenv
azure-core
rtclient
sounddevice
numpy
